Refactor car service methouds, add new packages for dto's
