+++ src/main/java/online/carsharing/controller/CarController.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import online.carsharing.dto.request.CreateCarRequestDto;
-import online.carsharing.dto.response.CarResponseDto;
-import online.carsharing.dto.update.CarUpdateDto;
+import online.carsharing.dto.request.car.CreateCarRequestDto;
+import online.carsharing.dto.response.car.CarResponseDto;
+import online.carsharing.dto.update.car.CarUpdateDto;
 import online.carsharing.service.CarService;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
