+++ src/main/java/online/carsharing/dto/update/car/CarUpdateDto.java
@@ -1,4 +1,4 @@
-package online.carsharing.dto.update;
+package online.carsharing.dto.update.car;
 
 import jakarta.validation.constraints.Min;
 import lombok.Data;
