+++ src/main/java/online/carsharing/mapper/CarMapper.java
@@ -1,8 +1,8 @@
 package online.carsharing.mapper;
 
 import online.carsharing.config.MapperConfig;
-import online.carsharing.dto.request.CreateCarRequestDto;
-import online.carsharing.dto.response.CarResponseDto;
+import online.carsharing.dto.request.car.CreateCarRequestDto;
+import online.carsharing.dto.response.car.CarResponseDto;
 import online.carsharing.entity.Car;
 import org.mapstruct.Mapper;
